backend/pyproject.toml
[tool.poetry]
name = "viyugam-backend"
version = "0.1.0"
description = "Backend for Viyugam Life OS"
authors = ["Anti-Gravity Architect <architect@viyugam.ai>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = "^0.30.0"
beanie = "^1.26.0"
motor = "^3.5.0"
pydantic = "^2.8.0"
pydantic-settings = "^2.4.0"
google-generativeai = "^0.7.0"
typer = "^0.12.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
black = "^24.4.0"
isort = "^5.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
